requirements.txt
django-debug-toolbar==1.6
django==1.9.4
django-extensions==1.6.1
django-filter==0.12.0
djangorestframework==3.3.2
psycopg2==2.6.1
django-autofixture==0.12.0
flake8==2.5.4
pygments==2.2.0
celery==4.1.0
redis==2.10.6
